rendu cpp 6 et 7

+++ Module_6/ex01/Serializer.cpp
@@ -27,10 +27,10 @@ Serializer::~Serializer()
 
 // Functions
 
-unsigned long Serializer::serialize(Data *ptr) {
-    return reinterpret_cast<unsigned long>(ptr);
+uintptr_t Serializer::serialize(Data *ptr) {
+    return reinterpret_cast<uintptr_t>(ptr);
 }
 
-Data *Serializer::deserialize(unsigned long raw) {
+Data *Serializer::deserialize(uintptr_t raw) {
     return reinterpret_cast<Data *>(raw);
 }
