+++ Module_6/ex01/Serializer.hpp
@@ -5,6 +5,7 @@
 #include "Data.hpp"
 # include <iostream>
 # include <cmath>
+#include <stdint.h>
 
 class Serializer
 {
@@ -14,8 +15,8 @@ class Serializer
 	// Exceptions
 
 	// Functions	
-    static unsigned long serialize(Data* ptr);
-    static Data* deserialize(unsigned long raw);
+    static uintptr_t serialize(Data* ptr);
+    static Data* deserialize(uintptr_t raw);
   private:
 	// Constructors and destructors
 	Serializer();
